Start of the work on parsing
A new header file, parsing.h, was created to help organize and modularize the parsing functionality of the application. This file was included in the cub.h header to make the new parsing functions widely available across the codebase. The added files were also reflected in the Makefile and sources.mk. This change aims to enhance code readability and maintainability."