Add drawing functionality to the project
Introduced new drawing functions that will aid in rendering visuals for this project. This includes creating a line, pixel placement, drawing a line on an image and handling different slopes for line drawing. Also made modifications to some existing implementations to make them compatible. This effort will provide a foundation for constructing visuals in later stages of the project.