Add libft library and its references in project files
Added the libft library as a submodule and its inclusion in the header file. The libft library contains reusable C function which is already used by many projects. Their addition in the Makefile allows these functions to be compiled and linked into the executable. This will make developing new features easier and faster because we have a robust set of tools readily available. The .gitmodules file ensures that this library is cloned whenever this repository is cloned. This makes it easier for new developers to set up their development environments.