Started Parsing logic with error checking
Added a new parsing.c file and started implementing the logic in it. It includes some basic error checks like verifying the number of arguments and validating the file extension. This is the start of building the parsing functionality for the application. In addition, the printf.h library was added to the cub.h file to support the printf function.