In the 'input_player.c' file, the conditions to alter the player's movements have been revised to correct the movement logic. This change was brought about due to incorrect movement directions - the keys mapped for each direction were leading to the opposite movements.
In the 'update_player.c' file, refined the player's angle and position update logic by adding more conditions and calculations, involving the player's current angle and speed. Also, debug logs were commented out to clean up the console output when the game is running. This change was done to make the character's movements more precise and natural, depending on the speed and orientation.