Refactor variable and structure names for clarity
Changes in the code mainly involve renaming variables and structures for the sake of clarity and better understanding. The variable `bpp` was renamed to `bits_per_pixel` throughout the code to make its purpose clearer.

Similarly, the structure `t_mlx_img` was renamed to `t_master_img`, defining the master image more explicitly. This change is reflected throughout the corresponding functions parameters and calls.

This refactoring was necessary to improve readability and maintainability of the code. With more descriptive names, other developers