Refactor error messages in parsing layer
Changed reference from `ERR` to `ERM`. This was done to maintain consistency in error naming convention across the project. Now, `ERM` is used in all instances to refer to error messages while `ERC` is used for error codes. This separation of error codes and messages improves readability and clarity in the code.