Added parsing sources to Makefile
Added new parsing files (parsing.c and get_textures.c) and corresponding header file (parsing.h) to the Makefile. This is done to include parsing functionality in the build, which is crucial for handling and processing the provided map for testing. Also improved "all" target for more elegant compilation process.