Remove outdated TODO and update resource cleanup
The outdated TODO comment was removed as the task it was referring to was completed earlier. Also, the code for cleanup of resources on unsuccessful parsing was updated. Instead of using mlx_destroy_display function, we now call free function, which properly deallocates the resources used by the parsing process.