Upgrade map and player rendering with raycasting
This commit implements fundamental rudimentary rendering for both the map and the player in the 3D game engine. The additions in hooks.c establish a simple 2D view of the map and player using the draw_map and draw_player functions in the game loop. The raycast.c file is updated to correct the handling of ray intersections beyond the map borders and to create precision in detecting wall hits. Consequently, render.c is adjusted to draw each slice of wall hit by the rays to construct a 3D projection for the player. This is a pivotal upgrade from the existing flat map view and sets the foundation for further enhancements of the game engine's graphical capabilities.