Revise hit data type in Ray struct and improve player controls
The data type for the "hit" variable in the Ray struct has been changed from int to bool for clearer and more efficient checks of hit statuses. A field for image structure has also been added in the Data struct for facilitating the manipulation and storage of image data.

In addition, player controls have been updated to standard WASD layout from the previous ZQSD layout for better game experience and conformation to popular gaming standards.