Fixed potential double-free in get_textures.c
Memory deallocation for the buffer in get_textures.c is rearranged to ensure that it is freed before breaking the loop. This change is aimed at preventing memory leaks in the code. The buff is now getting freed before each break and outside the core loop to cover all possible paths.