Add get_map_size function to read_map.c
Added get_map_size to read_map.c. This function evaluates the size of the provided map. This was done to allow for more accurate level design and gameplay calculations. Previously, map size was not considered. This function computes both dimensional values (size[0] and size[1]) of the map. Size[0] stands for the longest line in the map and size[1] is the total number of lines in the map.