Flip textures horizontally in the south and west walls
The aim of this commit is to correct an issue where the textures of the south and west walls were not being rendered correctly. After the textures are parsed and stored into `data->map.walls_text`, the function `flip_horizontaly` is now called to change the orientation of these textures. This adjustment ensures a correct viewing experience for the user.