Started work on mouse handling
A new mouse handling functionality has been included to improve user interaction capabilities. On a mouse-click event, the mouse goes into listening mode - its position is now centered on screen and the cursor gets hidden. This feature can be turned off by pressing 'r' key, which brings the cursor back to normal mode and positions it at the center of the screen.