Replace rewrite_img with fill_color in on_loop function
With this commit, we have replaced the `rewrite_img` function with `fill_color` function in the `on_loop` function. The `rewrite_img` function was previously being used to update the image. But in order to make the code more efficient and easier to manage, we have switched to using `fill_color`.

The `fill_color` function also takes in additional arguments for floor and ceiling color. These color values are now obtained from the map data, making it more flexible and reusable. This change should help in improving the performance and maintainability of the image updating process.

Lastly, there was a method "fill_color" declared in the file srcs/render/render.c which is now removed since it's definition is moved to a global header file. This makes the function reusable for the rest of the project.