"Updated bonus headers and improved minimap function
Renamed all the included headers in the cub_bonus.h file to include _bonus in their names, to better differentiate the bonus version from the regular version of the files.

Improved the minimap functionality in the minimap.c file, by adding a new put_rotated_point function to render the player's direction correctly on the minimap. The draw of the minimap was also moved to a separate minimap_draw function for better code organization.

Refactored some parts of the code and fixed minor bugs related to player movement and the minimap display."