Modified player angle movement and added render to main_bonus.c
In "mouse_hooks_bonus.c", the player's angular speed was enhanced and stored not into player.angle but to a new variable player.angle_mov for keep better track of this change in game mechanics. Also in "main_bonus.c", after setting the game map and the focus of the window, functions were added for filling color, render, image creation, and image pushing into the window to display these images instantly; this was done to improve user interaction and visual experience. Lastly, CFLAGS in Makefile was changed to remove  debugging and optimization flags and have a standard code compilation.