Fixed Memory Leaks in `hooks.c`
Added two blocks in the `hook_window()` function in `hooks.c` to ensure that images created during the program execution are properly destroyed/free'd before the program is terminated. This is to prevent any potential memory leak that may occur. By doing so, the program's efficiency can be maintained and the execution process would be much smoother.