Update Makefile to include `init_angles.c`
This commit introduces a new file, `init_angles.c`, into the Makefile for the build of the bonus version of cub3D. This was done to improve angle calculations by creating a new struct, `t_angle_data`, to store cosine and sine values associated with the player. This change allows minimization of computations and a simplification of the code structure. It also improves performance by replacing division operations with bitwise right shift operations. Additionally, file headers are updated and some code formatting adjustments are made where necessary.