Add ft_isspace function and improve whitespace handling
Implemented a new function, ft_isspace, that checks if a character is a whitespace character. This function has been added to char.h and has been used to improve the handling of whitespaces in the get_textures_utils.c and ft_atoi.c files. This was done to enhance the robustness of string parsing and whitespace stripping, ensuring that not just spaces but all whitespace characters are handled.
The ft_isspace function is also integrated into the Makefile to ensure smooth build and execution.