Update Makefile and refactor minimapBonus.c for bonus version
Updated the Makefile to use minimap_bonus.c instead of minimap.c when building the bonus version of the program. This was necessary to separate logic for the bonus and standard versions of the application.

Refactored minimap_bonus.c to fix a bug in the rendering of the minimap where the player's angle was not correctly converted from radians to degrees, resulting in incorrectly displayed player orientation on the minimap.