Reduce default window width in Makefile and force square window
In this commit, the default window width in the Makefile has been changed from 800 to 600. This modification will help support smaller screen sizes, ensuring appropriate display of content even on smaller monitors. Also, if the xrandr command is available, the window width is now set to match the display height, keeping the window size square-shaped.