"Optimize code by replacing division with bitwise shift"
This commit optimizes the performance of the code by replacing division operations with equivalent bitwise shift operations. Bitwise operations are generally faster than arithmetic operations such as division and this update is to leverage this performance benefit without altering the operational logic of the code. In addition, the definition of `CELL_SIZE` was changed from `64` to `512` and `CELL_SH` was defined as `9` to enable the bitwise shift operation.