"Refactor player movement and rotation speed constants"
This commit updates the player's movement and rotation speed. Previously, the PLAYER_SPEED constant was used for both linear and angular velocity of the player, which was not quite realistic. To make the game mechanics more accurate, PLAYER_SPEED is now applied solely for linear movement while a new constant, PLAYER_ROT_SPEED, is introduced specifically for rotation.

The values were also changed, PLAYER_SPEED from 3 to 25 and PLAYER_ROT_SPEED to 2, to better simulate the player's movement and rotation speed. This improves the gameplay by making the player's movement and rotation feel more lifelike.