Add re_bonus command to Makefile
Adding the re_bonus command to the Makefile for easier rebuilding of the bonus portion of the project. This changes simplifies the rebuilding of bonus changes without cleansing the whole project. This update should increase efficiency and save time for developers when they are working on the bonus part of the project.