Fixed FPS number display by switching it to use gettimeofday().
Previous implementation of the FPS counter was broken due to it using the clock() function to get the time elapsed since the start of the frame. This wasn't working as expected due tu the use of usleep() in the rendering loop.
By switching over to use gettimeofday(), we now have a more precise and reliable method to calculate and display the FPS.