Add FPS visibility toggle feature
A new boolean variable 'show_fps' has been added in the structure of types_bonus.h. This acts as a flag to toggle the visibility of frames per second (FPS) indicator on the screen. As a part of this change, render_bonus.c checks this flag, and if it's true, displays the FPS in the middle right of the screen. To control this new feature, a new key binding was set in hooks_bonus.c which sets the show_fps flag when F3 button is pressed. This improves the user experience by giving them an option to view/hide the FPS indicator as desired.