Correctly freeing every texture, avoiding double free with fallback textures.
Slightly change the exit routine of the program. Now, before freeing a texture for the wall or a sprite, the program will first check if the texture to free is the same as the one being pointing at by the default one. If that is the case, the program will skip the texture.
The default sprite are then freed later in the exit routine.