Optimize compilation flag and add norminette check
The Makefile has been updated to remove the -O2 optimization flag from the CFLAGS, allowing for more in-depth debugging. Additionally, the 'norminette' command has been added to the dependency list of the project name in the Makefile. This ensures that the norminette check, a coding standard verification tool, will be run before each compilation. This will help in maintaining coding standards across the project.