Remove 'norminette' dependency in Makefile
Norminette, a code normalizer, was being invoked as a dependency during code compilation. This was removed to ensure the software compiles successfully even if Norminette isn't installed or operational. This removal does not affect the functionality of the code but alleviates the unnecessary dependency. This adjustment will improve build process reliability.