Added a targeting reticle to increase weapon usability.
A targeting reticle was added to the weapon structure (types_bonus.h). Necessary changes to allocate and deallocate the reticle image were made in open_weapon_bonus.c and free_weapon_bonus.c respectively. The reticle is rendered on the screen in render_weapon_bonus.c. Additionally, the Makefile was updated to enable address sanitization which helps detect memory-related issues.