Add door checking in sprites parsing process
A "door" type has been added to the t_map struct in 'types_bonus.h' and parsing code for the door in 'sprites_parsing_bonus.c'. This update helps to distinguish doors from walls and sprites, and treat it differently in subsequent functionality. Also, planning for the function implementation is indicated with a 'TODO' comment.