Updated sprite parsing to handle duplicate information
Modified the sprite parsing method to return an error if duplicate sprite information is detected. A boolean flag has been added to the sprite struct to keep track of whether the sprite has been set. If a sprite has been set and the program attempts to set it again, an error code is returned and the memory allocated for the sprite path is freed appropriately. This change prevents possible duplication and unexpected behavior.