Normed gamepad related function (change update_gamepads return type from int to void)
Adjusted the return type of update_gamepads function from int to void in the gamepad_bonus files since the return value wasn't being used elsewhere. This alows the function to be less than 25 line. Fixed the format of variable 'i' in multiple files to stay consistent and improve readability. Added or maintained newlines at end of files to comply with POSIX standards.