"Added interactive doors feature in the game"
This commit adds an interactive doors feature in the game. In the game, we can now interact with the doors that can open and close using the 'E' key from the keyboard. The interaction with the doors is defined in the player input and player interact functions. New files for handling doors (doors_bonus.h, create_doors_map_bonus.c, destroy_doors_map_bonus.c) were added. It also added necessary checks in functions like ray_collides_bonus.c and update_player_bonus.c to handle player navigation through the doors. This interaction increases the complexity and interactivity of the game.